Main.class cleanup, bug fix
Die geometry Methoden haben jetzt eine eigene Klasse und die Gegner
buggen nicht mehr rum wenn das Fenster minimiert wird. (Musste die neue
Branch erstellen weil ich sonst die ganzen Merge conflicts beheben
müsste)
